Add fix for gulp error 'primordials is not defined'
This is not a direct fix but gives a file that can be used to fix the
error 'primordials is not defined' that can occur when building a CTO
plugin with gulp 3. The file then has to be copied into the directory
of the given plugin. Node modules have to be installed again also.